+++ package.json
@@ -12,8 +12,10 @@
   "dependencies": {
     "axios": "^0.18.0",
     "babel-runtime": "^6.0.0",
+    "better-scroll": "^1.13.2",
     "express": "^4.16.3",
     "fastclick": "^1.0.6",
+    "jsonp": "^0.2.1",
     "swiper": "^4.3.5",
     "vue": "^2.5.2",
     "vue-awesome-swiper": "^3.1.3",
