+++ src/common/stylus/variable.styl
@@ -10,7 +10,6 @@ $color-text = #fff
 $color-text-d = rgba(255, 255, 255, 0.3)
 $color-text-l = rgba(255, 255, 255, 0.5)
 $color-text-ll = rgba(255, 255, 255, 0.8)
-
 //字体定义规范
 $font-size-small-s = 10px
 $font-size-small = 12px
