+++ test.html
@@ -7,42 +7,83 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Document</title>
   <style>
-    .box{
-      text-align:center;
-      width:500px;
-      height:400px;
-      border:1px solid black;
-      margin:auto;
+    *{
+      margin:0px;
+      padding:0;
     }
-  .box::before{
-    content:"";
-    display:inline-block;
-    height:100%;
-    vertical-align:middle;
-  }
-    .box .content{
-      vertical-align:middle;
-      display:inline-block;
+    body{
+      background:#fff;
+      height:100vh;
+      position: relative;
+    }
+    .alertInfo{
+      padding-top:100px;
+    }
+    .alertInfo h2{
+      display:flex;
+      justify-content: center;
+      align-items: center;
+      color:#000;
+      font-size:20px;
+      margin-bottom:20px;
+      font-weight:400;
+    }
+    .alertInfo img{
+      width:50px;
+      height:50px;
+      margin-right:20px;
+    }
+    .alertInfo p{
+      font-size:18px;
+      color:#666;
+      text-align: center;
+      margin-bottom:50px;
+    }
+    .alertInfo span{
+      margin-right:5px;
+    }
+    .warningInfo{
+      position: absolute;
+      bottom:0;
+    }
+    .warningInfo p{
+      width:80%;
+      line-height:25px;
+      font-size:14px;
+      color:#666;
+      margin:auto;
     }
   </style>
 </head>
 
 <body>
-
-    <div class='box'>
-
-      <div class='content'>
-        日此精美猎奇的<a href="/">相册</a>
-        日此精美猎奇的
-        日此精美猎奇的
-        日此精美猎奇的
-      </div>
-
+    <div class='alertInfo'>
+      <h2><img src="successExchange@2x.png" alt="">已成功通知卖家收款！</h2>
+      <p>卖家确认收款后，您购买的USDT将从OTC平台释放到您充值的商家平台！</p>
+      <h2><span></span>S 后自动返回商户</h2>
     </div>
-  <script src='https://cdn.bootcss.com/jquery/1.10.2/jquery.min.js'></script>
-  <script>
-   
-  </script>
+    <div class='warningInfo'>
+        <p>交易须知：</p> 
+        <p>1、每次支付随机匹配的卖家不同，同一个卖家所使用的银行卡也可能不同，请按照每次所显示的付款信息打款，请勿直接打款到之前充值过的卡号，否则可能无法到账，造成的损失平台概不负责。</p> 
+        <p>2、转账时请勿填写包括数字货币、USDT、充值、美金、外汇等字样。否则可能导致卖家账户和您的账户被冻结，造成的损失平台概不负责。</p> 
+        <p>3、重要提示：作为独立公正的数字货币撮合平台，坚决反对任何机构和个人利用平台从事不符合法律的商业行为。如果您对交易有疑问，或有任何的投诉或建议，请email联系官方客服：
+        <p>4、根据国家相关法律规定，购买数字货币仅支持自然人之间的点对点转账交易</p> 
+        <p>5、卖家收到转账后您购买的USDT将从OTC平台释放到您充值的商家平台</p> 
+    </div>
 </body>
-
+<script src='http://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js'></script>
+<script>
+  var n=6;
+  function timer(){
+    n--;
+    
+    if(n<0){
+      window.history.back();
+      return
+    }
+    $('.alertInfo span').text(n);
+    setTimeout(timer,1000)
+  }
+  timer();
+</script>
 </html>
